Create a `strict-contains` matching mode for a more obvious sort order.
The `contains` matching mode is not just a `contains` mode. It also implies a change to the sort
order. Namely, in `contains` mode if one of the options is prefixed by the filter it will be
preferred over other options, even if they are more frequent in the cache. For example, if I had
the following options:

```
Filter: cha

Characters -- Used never
Google Chat -- Used all the time
```

`Characters` will be sorted before `Google Chat` because `Characters` happens to start with `cha`.
This leads to the wrong assumption that the cache is somehow broken (since something infrequent got
sorted first).

The `strict-contains` mode **only** cares about containment. Therefore, as long as two options both
contain the filter, they are considered equal, and will defer to the sort order (usually the
cache).

We could instead change the `contains` mode to remove this prefix checking thing. However, this may
be a desirable feature to someone so I chose to just create a separate matching mode instead.